requirements.txt
pyTelegramBotAPI==4.23.0
pymongo==4.10.1
python-dotenv==1.0.1
